Extract KFunction and KCaptureFunction into a separate module
