+++ core/runtime/src/vm.rs
@@ -2,11 +2,8 @@ use crate::{
     core_lib::CoreLib,
     error::RuntimeErrorType,
     prelude::*,
-    types::{
-        meta_id_to_key,
-        value::{KFunction, RegisterSlice},
-    },
-    DefaultStderr, DefaultStdin, DefaultStdout, KCaptureFunction, Result,
+    types::{meta_id_to_key, value::RegisterSlice},
+    DefaultStderr, DefaultStdin, DefaultStdout, KCaptureFunction, KFunction, Result,
 };
 use koto_bytecode::{Chunk, Instruction, InstructionReader, Loader, TypeId};
 use koto_parser::{ConstantIndex, MetaKeyId};
